+++ steward/discovery/discovery-owl.js
@@ -1,11 +1,8 @@
-// Thing Sensor Reporting Protocol
+// Owl Intution monitoring
 
 var dgram       = require('dgram')
-  , util        = require('util')
   , xml         = require('xml2json')
-  , trsp		= require('./discovery-tsrp')
-  , devices     = require('./../core/device')
-  , things      = require('./../api/api-manage-thing')
+  , trsp        = require('./discovery-tsrp')
   , utility     = require('./../core/utility')
   ;
 
@@ -13,202 +10,205 @@ var dgram       = require('dgram')
 var logger = utility.logger('discovery');
 
 exports.start = function() {
- 	var LOCAL_BROADCAST_HOST = '224.192.32.19';
-	var LOCAL_BROADCAST_PORT = 22600;
-	var bootTime = process.hrtime()[0] * 1000 + process.hrtime()[1] / 1000;
-	var requestID = 0;
-	
-	var socket = dgram.createSocket('udp4');
-	socket.bind(LOCAL_BROADCAST_PORT, function() {
-		socket.addMembership(LOCAL_BROADCAST_HOST);
-	});
-	
-	socket.on('message', function(message, rinfo) {
-    	var report;
+  var LOCAL_BROADCAST_HOST = '224.192.32.19';
+  var LOCAL_BROADCAST_PORT = 22600;
+  var bootTime = process.hrtime()[0] * 1000 + process.hrtime()[1] / 1000;
+  var requestID = 0;
 
-    	try { 
-			var json = xml.toJson( message );
-			var buff = JSON.parse( json );
-			var currTime = process.hrtime()[0] * 1000 + process.hrtime()[1] / 1000;
-			requestID = requestID+1;
-			
-			if ( buff.electricity ) {			
-				logger.info('Recieved OWL electricity packet.');
-				var battery = buff.electricity.battery.level.substring(0, buff.electricity.battery.level.length - 1);
-			
-				var channel0, channel1, channel2 = null;
-				buff.electricity.chan.forEach(function(entry) {
-				    if( entry.id === 0 ) {
-						channel0 = [{'current':entry.curr.$t,'units':entry.curr.units},
-									{'day':entry.day.$t,'units':entry.day.units}];
-					}
-					if( entry.id === 1 ) {
-						channel1 = [{'current':entry.curr.$t,'units':entry.curr.units},
-									{'day':entry.day.$t,'units':entry.day.units}];
-					}
-					if( entry.id === 2 ) {
-						channel2 = [{'current':entry.curr.$t,'units':entry.curr.units},
-									{'day':entry.day.$t,'units':entry.day.units}];
-					}
-				});
-				
-				report = {"path":"/api/v1/thing/reporting",
- 		  		          "requestID":requestID.toString(),
-				   		  "things":{
-				      		"/device/sensor/owl/electricity":{
-				         		"prototype":{
-				            		"device":{
-				               			"name":"OWL Intuition-e",
-				               			"maker":"2 Save Energy Ltd"
-				            		},
-				            		"name":true,
-				            		"status":[ "present", "absent", "recent"],
-				            		"properties":{ "rssi":"dB", "lqi":"", "battery":"%", "channel0": { "currentUsage":"W", "dailyUsage":"Wh" },
-																						 "channel1": { "currentUsage":"W", "dailyUsage":"Wh" },
-																						 "channel2": { "currentUsage":"W", "dailyUsage":"Wh" }}
-				         		},
-				         		"instances":[{
-				               		"name":"OWL Intuition-e",
-				               		"status":"present",
-				               		"unit":{
-				                  		"serial":buff.electricity.id,
-				                  		"udn":"195a42b0-ef6b-11e2-99d0-UID"+buff.electricity.id+"-owl-electricity-monitor"
-				               		},
-				               		"info":{
-				                  		"rssi":buff.electricity.signal.rssi,
-										"lqi":buff.electricity.signal.lqi,
-										"battery":battery,
-				                  		"channel0":{"currentUsage":channel0[0].current, "dailyUsage":channel0[1].day},
-				                  		"channel1":{"currentUsage":channel1[0].current, "dailyUsage":channel1[1].day},
-				                  		"channel2":{"currentUsage":channel2[0].current, "dailyUsage":channel2[1].day},
-				               		},
-				               		"uptime":currTime-bootTime
-				            	}]
-							}
-				   		}
-						};			
-			
-			} else if ( buff.heating ) {
-				logger.info('Recieved OWL heating packet.');
-				var battery = buff.heating.battery.level.substring(0, buff.heating.battery.level.length - 2);
-				report = {"path":"/api/v1/thing/reporting",
- 		  		  		  "requestID":requestID.toString(),
-				   		  "things":{
-				      		"/device/climate/owl/monitor":{
-				         		"prototype":{
-				            		"device":{
-				               			"name":"OWL Intuition-c",
-				               			"maker":"2 Save Energy Ltd"
-				            		},
-				            		"name":true,
-				            		"status":[ "present", "absent", "recent"],
-				            		"properties":{ "rssi":"dB", "lqi":"", "battery":"mV", "temperature":"celsius", "goalTemperature":"celsius",
-				                                   "until":"","zone":"" }
-				         		},
-				         		"instances":[{
-				               		"name":"OWL Intuition-c",
-				               		"status":"present",
-				               		"unit":{
-				                  		"serial":buff.heating.id,
-				                  		"udn":"195a42b0-ef6b-11e2-99d0-UID"+buff.heating.id+"-owl-thermostat"
-				               		},
-				               		"info":{
-				                  		"rssi":buff.heating.signal.rssi,
-										"lqi":buff.heating.signal.lqi,
-										"battery":battery,
-				                  		"temperature":buff.heating.temperature.current,
-				                  		"goalTemperature":buff.heating.temperature.required,
-										"until":buff.heating.temperature.until,
-										"zone":buff.heating.temperature.zone.toString()
-				               		},
-				               		"uptime":currTime-bootTime
-				            	}]
-							}
-				   		}
-						};
-		
-		 	} else if ( buff.weather ) {
-				logger.info('Recieved OWL weather packet.');
-				report = {"path":"/api/v1/thing/reporting",
-		   		  		  "requestID":requestID.toString(),
-				   		  "things":{
-				      		"/device/climate/owl/sensor":{
-				         		"prototype":{
-				            		"device":{
-				               			"name":"Network OWL",
-				               			"maker":"2 Save Energy Ltd"
-				            		},
-				            		"name":true,
-				            		"status":[ "present", "absent", "recent"],
-				            		"properties":{ "code":"", "temperature":"celsius", "text":"" }
-				         		},
-				         		"instances":[{
-				               		"name":"OWL Weather",
-				               		"status":"present",
-				               		"unit":{
-				                  		"serial":buff.weather.id,
-				                  		"udn":"195a42b0-ef6b-11e2-99d0-UID"+buff.weather.id+"-owl-weather"
-				               		},
-				               		"info":{
-				                  		"code":buff.weather.code,
-				                  		"temperature":buff.weather.temperature,
-				                  		"text":buff.weather.text
-				               		},
-				               		"uptime":currTime-bootTime
-				            	}]
-							}
-				   		}
-						};
-		
-			} else if ( buff.solar ) {
-				logger.info('Recieved OWL solar packet.');
-				report = {"path":"/api/v1/thing/reporting",
-				  		  "requestID":requestID.toString(),
-				   		  "things":{
-				      		"/device/sensor/owl/solarpanels":{
-				         		"prototype":{
-				            		"device":{
-				               			"name":"OWL Intuition-pv",
-				               			"maker":"2 Save Energy Ltd"
-				            		},
-				            		"name":true,
-				            		"status":[ "present", "absent", "recent"],
-				            		"properties":{ "generating":"W", "exporting":"W" }
-				         		},
-				         		"instances":[{
-				               		"name":"OWL Intuition-pv",
-				               		"status":"present",
-				               		"unit":{
-				                  		"serial":buff.solar.id,
-				                  		"udn":"195a42b0-ef6b-11e2-99d0-UID"+buff.solar.id+"-owl-solarpanel"
-				               		},
-				               		"info":{
-				                  		"generating":buff.solar.current.generating,
-				                  		"exporting":buff.solar.current.exporting
-				               		},
-				               		"uptime":currTime-bootTime
-				            	}]
-							}
-				   		}
-						};		
-			}
-			
-			//console.log(util.inspect(report, false, null));
-			trsp.handle(report, rinfo.address, 'udp ' + rinfo.address + ' ' + rinfo.port + ' ' + report.path);
-	    } catch(ex) {
-		    logger.error('reporting', { event: 'parsing', diagnostic: ex.message });
-      		return;
-    	}
-  	});
+  var socket = dgram.createSocket('udp4');
 
-  	socket.on('listening', function() {
-    	var address = this.address();
-		logger.info('OWL driver listening on udp://*:' + address.port);
-  	});
+  socket.bind(LOCAL_BROADCAST_PORT, function() {
+      socket.addMembership(LOCAL_BROADCAST_HOST);
+  });
 
-	socket.on('error', function(err) {
-    	logger.error('reporting', { event: 'socket', diagnostic: err.message });
-  	})
+  socket.on('message', function(message, rinfo) {
+      var report;
 
-	
+      try {
+        var json = xml.toJson( message );
+        var buff = JSON.parse( json );
+        var currTime = process.hrtime()[0] * 1000 + process.hrtime()[1] / 1000;
+        requestID = requestID+1;
+
+        if ( buff.electricity ) {
+          logger.info('Received OWL electricity packet.');
+
+          var channels = [{}, {}, {}];
+          buff.electricity.chan.forEach(function(entry) {
+            channels[entry.id] = { 'current': entry.curr.$t, 'daily': entry.day.$t };
+          });
+
+// not reporting because we don't know the units: lqi
+          report = { "path"                             : "/api/v1/thing/reporting"
+                   , "requestID"                        : requestID.toString()
+                   , "things":
+                     { "/device/sensor/owl/electricity" :
+                       { "prototype"                    :
+                         { "device"                     :
+                           { "name"                     : "OWL Intuition-e"
+                           , "maker"                    : "2 Save Energy Ltd"
+                           }
+                         , "name"                       : true
+                         , "status"                     : [ "present", "absent", "recent"]
+                         , "properties":
+                           { "rssi"                     : "s8"
+                           , "batteryLevel"             : "percentage"
+                           , "currentUsage"             : "watts"
+                           , "dailyUsage"               : "watt-hours"
+                           }
+                         }
+                       , "instances"                    :
+                         [ { "name"                     : "OWL Intuition-e"
+                           , "status"                   : "present"
+                           , "unit"                     :
+                             { "serial"                 : buff.electricity.id
+                             , "udn"                    : "195a42b0-ef6b-11e2-99d0-UID"+buff.electricity.id+"-owl-electricity"
+                             }
+                           , "info"                     :
+                             { "rssi"                   : buff.electricity.signal.rssi
+                             , "batteryLevel"           : parseFloat(buff.electricity.battery.level)
+                             , "currentUsage"           : [ channels[0].current, channels[1].current, channels[2].current ]
+                             , "dailyUsage"             : [ channels[0].daily,   channels[1].daily,   channels[2].daily   ]
+                             }
+                           , "uptime"                   : currTime-bootTime
+                           }
+                         ]
+                       }
+                     }
+                   };
+
+        } else if ( buff.heating ) {
+          logger.info('Received OWL heating packet.');
+
+// not reporting because we don't know the units: lqi; or we don't care: until/zone
+          report = { "path"                             : "/api/v1/thing/reporting"
+                   , "requestID"                        : requestID.toString()
+                   , "things":
+                     { "/device/climate/owl/monitor"    :
+                       { "prototype"                    :
+                         { "device"                     :
+                           { "name"                     : "OWL Intuition-c"
+                           , "maker"                    : "2 Save Energy Ltd"
+                           }
+                         , "name"                       : true
+                         , "status"                     : [ "present", "absent", "recent"]
+                         , "properties":
+                           { "rssi"                     : "s8"
+                           , "battery"                  : "volts"
+                           , "temperature"              : "celsius"
+                           , "goalTemperature"          : "celsius"
+                           }
+                         }
+                       , "instances"                    :
+                         [ { "name"                     : "OWL Intuition-c"
+                           , "status"                   : "present"
+                           , "unit"                     :
+                             { "serial"                 : buff.heating.id
+                             , "udn"                    : "195a42b0-ef6b-11e2-99d0-UID"+buff.heating.id+"-owl-thermostat"
+                             }
+                           , "info"                     :
+                             { "rssi"                   : buff.heating.signal.rssi
+                             , "battery"                : parseFloat(buff.heating.battery.level.toString()) / 1000.0
+                             , "temperature"            : buff.heating.temperature.current
+                             , "goalTemperature"        : buff.heating.temperature.required
+                             }
+                           , "uptime"                   : currTime-bootTime
+                           }
+                         ]
+                       }
+                     }
+                   };
+
+        } else if ( buff.weather ) {
+          logger.info('Received OWL weather packet.');
+
+          report = { "path"                             : "/api/v1/thing/reporting"
+                   , "requestID"                        : requestID.toString()
+                   , "things":
+                     { "/device/climate/owl/sensor"    :
+                       { "prototype"                    :
+                         { "device"                     :
+                           { "name"                     : "Network OWL"
+                           , "maker"                    : "2 Save Energy Ltd"
+                           }
+                         , "name"                       : true
+                         , "status"                     : [ "present", "absent", "recent"]
+                         , "properties":
+                           { "code"                     : true
+                           , "temperature"              : "celsius"
+                           , "text"                     : true
+                           }
+                         }
+                       , "instances"                    :
+                         [ { "name"                     : "Network OWL"
+                           , "status"                   : "present"
+                           , "unit"                     :
+                             { "serial"                 : buff.weather.id
+                             , "udn"                    : "195a42b0-ef6b-11e2-99d0-UID"+buff.weather.id+"-owl-weather"
+                             }
+                           , "info"                     :
+                             { "code"                   : buff.weather.code
+                             , "temperature"            : buff.weather.temperature
+                             , "text"                   : buff.weather.text
+                             }
+                           , "uptime"                   : currTime-bootTime
+                           }
+                         ]
+                       }
+                     }
+                   };
+
+        } else if ( buff.solar ) {
+          logger.info('Received OWL solar packet.');
+
+          report = { "path"                             : "/api/v1/thing/reporting"
+                   , "requestID"                        : requestID.toString()
+                   , "things":
+                     { "/device/sensor/owl/solarpanels" :
+                       { "prototype"                    :
+                         { "device"                     :
+                           { "name"                     : "OWL Intuition-pv"
+                           , "maker"                    : "2 Save Energy Ltd"
+                           }
+                         , "name"                       : true
+                         , "status"                     : [ "present", "absent", "recent"]
+                         , "properties":
+                           { "generating"               : "watts"
+                           , "exporting"                : "watts"
+                           }
+                         }
+                       , "instances"                    :
+                         [ { "name"                     : "OWL Intuition-pv"
+                           , "status"                   : "present"
+                           , "unit"                     :
+                             { "serial"                 : buff.solar.id
+                             , "udn"                    : "195a42b0-ef6b-11e2-99d0-UID"+buff.solar.id+"-owl-solarpanel"
+                             }
+                           , "info"                     :
+                             { "generating"             : buff.solar.current.generating
+                             , "exporting"              : buff.solar.current.exporting
+                             }
+                           , "uptime"                   : currTime-bootTime
+                           }
+                         ]
+                       }
+                     }
+                   };
+
+        }
+
+        //console.log(util.inspect(report, false, null));
+        trsp.handle(report, rinfo.address, 'udp ' + rinfo.address + ' ' + rinfo.port + ' ' + report.path);
+      } catch(ex) { return logger.error('reporting', { event: 'parsing', diagnostic: ex.message }); }
+    });
+
+    socket.on('listening', function() {
+      var address = this.address();
+      logger.info('OWL driver listening on udp://*:' + address.port);
+    });
+
+    socket.on('error', function(err) {
+      logger.error('reporting', { event: 'socket', diagnostic: err.message });
+    });
+
 };